+++ plugins/source/aws/client/tableoptions/cloudwatch.go
@@ -7,47 +7,28 @@ import (
 	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/service/cloudwatch"
 	"github.com/cloudquery/plugin-sdk/v4/caser"
+	"github.com/invopop/jsonschema"
 )
 
-type CloudwatchGetMetricStatisticsInput struct {
-	cloudwatch.GetMetricStatisticsInput
-}
-
-// UnmarshalJSON implements the json.Unmarshaler interface for the CloudwatchGetMetricStatisticsInput type.
-// It is the same as default, but allows the use of underscore in the JSON field names.
-func (c *CloudwatchGetMetricStatisticsInput) UnmarshalJSON(data []byte) error {
-	m := map[string]any{}
-	if err := json.Unmarshal(data, &m); err != nil {
-		return err
+type (
+	CloudwatchMetrics []CloudwatchMetric
+	CloudwatchMetric  struct {
+		ListMetricsOpts         CloudwatchListMetricsInput           `json:"list_metrics,omitempty"`
+		GetMetricStatisticsOpts []CloudwatchGetMetricStatisticsInput `json:"get_metric_statistics,omitempty"`
 	}
-	csr := caser.New()
-	changeCaseForObject(m, csr.ToPascal)
-	b, _ := json.Marshal(m)
-	return json.Unmarshal(b, &c.GetMetricStatisticsInput)
-}
+)
 
-type CloudwatchMetric struct {
-	ListMetricsOpts         CloudwatchListMetricsInput           `json:"list_metrics,omitempty"`
-	GetMetricStatisticsOpts []CloudwatchGetMetricStatisticsInput `json:"get_metric_statistics,omitempty"`
-}
-
-type CloudwatchMetrics []CloudwatchMetric
-
-type CloudwatchListMetricsInput struct {
-	cloudwatch.ListMetricsInput
-}
-
-// UnmarshalJSON implements the json.Unmarshaler interface for the CloudwatchListMetricsInput type.
-// It is the same as default, but allows the use of underscore in the JSON field names.
-func (c *CloudwatchListMetricsInput) UnmarshalJSON(data []byte) error {
-	m := map[string]any{}
-	if err := json.Unmarshal(data, &m); err != nil {
-		return err
+func (c CloudwatchMetrics) Validate() error {
+	for _, m := range c {
+		if err := m.Validate(); err != nil {
+			return err
+		}
 	}
-	csr := caser.New()
-	changeCaseForObject(m, csr.ToPascal)
-	b, _ := json.Marshal(m)
-	return json.Unmarshal(b, &c.ListMetricsInput)
+	return nil
+}
+
+func (c *CloudwatchMetric) Validate() error {
+	return errors.Join(c.validateListMetricsOpts(), c.validateGetMetricStatisticsOpts())
 }
 
 func (c *CloudwatchMetric) validateListMetricsOpts() error {
@@ -72,21 +53,50 @@ func (c CloudwatchMetric) validateGetMetricStatisticsOpts() error {
 	return nil
 }
 
-func (c *CloudwatchMetric) Validate() error {
-	if err := c.validateListMetricsOpts(); err != nil {
-		return err
-	}
-	return c.validateGetMetricStatisticsOpts()
+type CloudwatchGetMetricStatisticsInput struct {
+	cloudwatch.GetMetricStatisticsInput
 }
 
-func (c CloudwatchMetrics) Validate() error {
-	for _, m := range c {
-		if err := m.validateListMetricsOpts(); err != nil {
-			return err
-		}
-		if err := m.validateGetMetricStatisticsOpts(); err != nil {
-			return err
-		}
+// JSONSchemaExtend is required to remove `namespace`, `metric_name` & `dimensions`.
+// We use value receiver because of https://github.com/invopop/jsonschema/issues/102
+func (CloudwatchGetMetricStatisticsInput) JSONSchemaExtend(sc *jsonschema.Schema) {
+	sc.Properties.Delete("Namespace")  // we don't allow `namespace`
+	sc.Properties.Delete("MetricName") // we don't allow `metric_name`
+	sc.Properties.Delete("Dimensions") // we don't allow `dimensions`
+}
+
+// UnmarshalJSON implements the json.Unmarshaler interface for the CloudwatchGetMetricStatisticsInput type.
+// It is the same as default, but allows the use of underscore in the JSON field names.
+func (c *CloudwatchGetMetricStatisticsInput) UnmarshalJSON(data []byte) error {
+	m := map[string]any{}
+	if err := json.Unmarshal(data, &m); err != nil {
+		return err
 	}
-	return nil
+	csr := caser.New()
+	changeCaseForObject(m, csr.ToPascal)
+	b, _ := json.Marshal(m)
+	return json.Unmarshal(b, &c.GetMetricStatisticsInput)
+}
+
+type CloudwatchListMetricsInput struct {
+	cloudwatch.ListMetricsInput
+}
+
+// UnmarshalJSON implements the json.Unmarshaler interface for the CloudwatchListMetricsInput type.
+// It is the same as default, but allows the use of underscore in the JSON field names.
+func (c *CloudwatchListMetricsInput) UnmarshalJSON(data []byte) error {
+	m := map[string]any{}
+	if err := json.Unmarshal(data, &m); err != nil {
+		return err
+	}
+	csr := caser.New()
+	changeCaseForObject(m, csr.ToPascal)
+	b, _ := json.Marshal(m)
+	return json.Unmarshal(b, &c.ListMetricsInput)
+}
+
+// JSONSchemaExtend is required to remove `next_token`.
+// We use value receiver because of https://github.com/invopop/jsonschema/issues/102
+func (CloudwatchListMetricsInput) JSONSchemaExtend(sc *jsonschema.Schema) {
+	sc.Properties.Delete("NextToken") // we don't allow `next_token`
 }
